requirements.txt
Flask
joblib
numpy
google-generativeai
